requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
attrs==24.2.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
loguru==0.7.2
multidict==6.0.5
numpy==2.0.1
pandas==2.2.2
psycopg2-binary==2.9.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
